requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0  # 连接MySQL（如需PostgreSQL可换psycopg2）
python-dotenv==1.0.0  # 环境变量管理
flask-jwt-extended==4.6.0
boto3==1.34.0